LayoutTests/fast/css-grid-layout/named-grid-lines-with-named-grid-areas-get-set-expected.txt

PASS getComputedStyle(GridUniqueColumnB, '').getPropertyValue('grid-column') is "b / b"
PASS getComputedStyle(GridUniqueColumnB, '').getPropertyValue('grid-column-start') is "b"
PASS getComputedStyle(GridUniqueColumnB, '').getPropertyValue('grid-column-end') is "b"
PASS getComputedStyle(GridUniqueRowE, '').getPropertyValue('grid-column') is "auto / auto"
PASS getComputedStyle(GridUniqueRowE, '').getPropertyValue('grid-column-start') is "auto"
PASS getComputedStyle(GridUniqueRowE, '').getPropertyValue('grid-column-end') is "auto"
PASS getComputedStyle(GridUniqueColumnBStart, '').getPropertyValue('grid-column') is "b-start / b-start"
PASS getComputedStyle(GridUniqueColumnBStart, '').getPropertyValue('grid-column-start') is "b-start"
PASS getComputedStyle(GridUniqueColumnBStart, '').getPropertyValue('grid-column-end') is "b-start"
PASS getComputedStyle(GridUniqueRowGStart, '').getPropertyValue('grid-column') is "auto / auto"
PASS getComputedStyle(GridUniqueRowGStart, '').getPropertyValue('grid-column-start') is "auto"
PASS getComputedStyle(GridUniqueRowGStart, '').getPropertyValue('grid-column-end') is "auto"
PASS getComputedStyle(GridUniqueColumnBD, '').getPropertyValue('grid-column') is "b / d"
PASS getComputedStyle(GridUniqueColumnBD, '').getPropertyValue('grid-column-start') is "b"
PASS getComputedStyle(GridUniqueColumnBD, '').getPropertyValue('grid-column-end') is "d"
PASS getComputedStyle(GridUniqueRowGH, '').getPropertyValue('grid-column') is "auto / auto"
PASS getComputedStyle(GridUniqueRowGH, '').getPropertyValue('grid-column-start') is "auto"
PASS getComputedStyle(GridUniqueRowGH, '').getPropertyValue('grid-column-end') is "auto"
PASS getComputedStyle(GridUniqueRowGH, '').getPropertyValue('grid-row') is "g / h"
PASS getComputedStyle(GridUniqueRowGH, '').getPropertyValue('grid-row-start') is "g"
PASS getComputedStyle(GridUniqueRowGH, '').getPropertyValue('grid-row-end') is "h"
PASS getComputedStyle(GridUniqueColumnCRowF, '').getPropertyValue('grid-column') is "c / c"
PASS getComputedStyle(GridUniqueColumnCRowF, '').getPropertyValue('grid-column-start') is "c"
PASS getComputedStyle(GridUniqueColumnCRowF, '').getPropertyValue('grid-column-end') is "c"
PASS getComputedStyle(GridUniqueColumnBInteger, '').getPropertyValue('grid-column') is "b / 4"
PASS getComputedStyle(GridUniqueColumnBInteger, '').getPropertyValue('grid-column-start') is "b"
PASS getComputedStyle(GridUniqueColumnBInteger, '').getPropertyValue('grid-column-end') is "4"
PASS getComputedStyle(GridUniqueRowIntegerH, '').getPropertyValue('grid-column') is "auto / auto"
PASS getComputedStyle(GridUniqueRowIntegerH, '').getPropertyValue('grid-column-start') is "auto"
PASS getComputedStyle(GridUniqueRowIntegerH, '').getPropertyValue('grid-column-end') is "auto"
PASS getComputedStyle(GridUniqueRowIntegerH, '').getPropertyValue('grid-row') is "3 / h"
PASS getComputedStyle(GridUniqueRowIntegerH, '').getPropertyValue('grid-row-start') is "3"
PASS getComputedStyle(GridUniqueRowIntegerH, '').getPropertyValue('grid-row-end') is "h"
PASS getComputedStyle(GridUniqueColumnIntegerRowG, '').getPropertyValue('grid-column') is "2 / auto"
PASS getComputedStyle(GridUniqueColumnIntegerRowG, '').getPropertyValue('grid-column-start') is "2"
PASS getComputedStyle(GridUniqueColumnIntegerRowG, '').getPropertyValue('grid-column-end') is "auto"
PASS getComputedStyle(GridUniqueColumnARowInteger, '').getPropertyValue('grid-column') is "a / a"
PASS getComputedStyle(GridUniqueColumnARowInteger, '').getPropertyValue('grid-column-start') is "a"
PASS getComputedStyle(GridUniqueColumnARowInteger, '').getPropertyValue('grid-column-end') is "a"
PASS getComputedStyle(GridUniqueColumnCStart, '').getPropertyValue('grid-column') is "c-start / c-start"
PASS getComputedStyle(GridUniqueColumnCStart, '').getPropertyValue('grid-column-start') is "c-start"
PASS getComputedStyle(GridUniqueColumnCStart, '').getPropertyValue('grid-column-end') is "c-start"
PASS getComputedStyle(GridUniqueRowFStart, '').getPropertyValue('grid-column') is "auto / auto"
PASS getComputedStyle(GridUniqueRowFStart, '').getPropertyValue('grid-column-start') is "auto"
PASS getComputedStyle(GridUniqueRowFStart, '').getPropertyValue('grid-column-end') is "auto"
PASS getComputedStyle(GridUniqueColumnCStartRowFEnd, '').getPropertyValue('grid-column') is "c-start / c-start"
PASS getComputedStyle(GridUniqueColumnCStartRowFEnd, '').getPropertyValue('grid-column-start') is "c-start"
PASS getComputedStyle(GridUniqueColumnCStartRowFEnd, '').getPropertyValue('grid-column-end') is "c-start"
PASS getComputedStyle(GridUniqueColumnBEndRowHStart, '').getPropertyValue('grid-column') is "b-end / b-end"
PASS getComputedStyle(GridUniqueColumnBEndRowHStart, '').getPropertyValue('grid-column-start') is "b-end"
PASS getComputedStyle(GridUniqueColumnBEndRowHStart, '').getPropertyValue('grid-column-end') is "b-end"
PASS getComputedStyle(GridAreasNoNamesColumnAStartRowDStart, '').getPropertyValue('grid-column') is "a-start / a-start"
PASS getComputedStyle(GridAreasNoNamesColumnAStartRowDStart, '').getPropertyValue('grid-column-start') is "a-start"
PASS getComputedStyle(GridAreasNoNamesColumnAStartRowDStart, '').getPropertyValue('grid-column-end') is "a-start"
PASS getComputedStyle(GridAreasNoNamesColumnAStartRowD, '').getPropertyValue('grid-column') is "a-start / a-start"
PASS getComputedStyle(GridAreasNoNamesColumnAStartRowD, '').getPropertyValue('grid-column-start') is "a-start"
PASS getComputedStyle(GridAreasNoNamesColumnAStartRowD, '').getPropertyValue('grid-column-end') is "a-start"
PASS getComputedStyle(GridAreasNoNamesColumnDStartRowCStart, '').getPropertyValue('grid-column') is "d-start / d-start"
PASS getComputedStyle(GridAreasNoNamesColumnDStartRowCStart, '').getPropertyValue('grid-column-start') is "d-start"
PASS getComputedStyle(GridAreasNoNamesColumnDStartRowCStart, '').getPropertyValue('grid-column-end') is "d-start"
PASS getComputedStyle(GridAreasNoNamesColumnD, '').getPropertyValue('grid-column') is "d / d"
PASS getComputedStyle(GridAreasNoNamesColumnD, '').getPropertyValue('grid-column-start') is "d"
PASS getComputedStyle(GridAreasNoNamesColumnD, '').getPropertyValue('grid-column-end') is "d"
PASS getComputedStyle(GridAreasNoNamesRowD, '').getPropertyValue('grid-column') is "auto / auto"
PASS getComputedStyle(GridAreasNoNamesRowD, '').getPropertyValue('grid-column-start') is "auto"
PASS getComputedStyle(GridAreasNoNamesRowD, '').getPropertyValue('grid-column-end') is "auto"
PASS getComputedStyle(GridAreasNoNamesColumnC, '').getPropertyValue('grid-column') is "c / c"
PASS getComputedStyle(GridAreasNoNamesColumnC, '').getPropertyValue('grid-column-start') is "c"
PASS getComputedStyle(GridAreasNoNamesColumnC, '').getPropertyValue('grid-column-end') is "c"
PASS getComputedStyle(GridAreasNoNamesRowC, '').getPropertyValue('grid-column') is "auto / auto"
PASS getComputedStyle(GridAreasNoNamesRowC, '').getPropertyValue('grid-column-start') is "auto"
PASS getComputedStyle(GridAreasNoNamesRowC, '').getPropertyValue('grid-column-end') is "auto"
PASS getComputedStyle(GridAreasNoNamesColumnA, '').getPropertyValue('grid-column') is "a / a"
PASS getComputedStyle(GridAreasNoNamesColumnA, '').getPropertyValue('grid-column-start') is "a"
PASS getComputedStyle(GridAreasNoNamesColumnA, '').getPropertyValue('grid-column-end') is "a"
PASS getComputedStyle(GridAreasNoNamesRowA, '').getPropertyValue('grid-column') is "auto / auto"
PASS getComputedStyle(GridAreasNoNamesRowA, '').getPropertyValue('grid-column-start') is "auto"
PASS getComputedStyle(GridAreasNoNamesRowA, '').getPropertyValue('grid-column-end') is "auto"
PASS getComputedStyle(GridAreasNoNamesColumnARowA, '').getPropertyValue('grid-column') is "a / a"
PASS getComputedStyle(GridAreasNoNamesColumnARowA, '').getPropertyValue('grid-column-start') is "a"
PASS getComputedStyle(GridAreasNoNamesColumnARowA, '').getPropertyValue('grid-column-end') is "a"
PASS getComputedStyle(GridAreasNamedLineAfterColumnD, '').getPropertyValue('grid-column') is "d / d"
PASS getComputedStyle(GridAreasNamedLineAfterColumnD, '').getPropertyValue('grid-column-start') is "d"
PASS getComputedStyle(GridAreasNamedLineAfterColumnD, '').getPropertyValue('grid-column-end') is "d"
PASS getComputedStyle(GridAreasNamedLineAfterRowD, '').getPropertyValue('grid-column') is "auto / auto"
PASS getComputedStyle(GridAreasNamedLineAfterRowD, '').getPropertyValue('grid-column-start') is "auto"
PASS getComputedStyle(GridAreasNamedLineAfterRowD, '').getPropertyValue('grid-column-end') is "auto"
PASS getComputedStyle(GridAreasNamedLineAfterColumnC, '').getPropertyValue('grid-column') is "c / c"
PASS getComputedStyle(GridAreasNamedLineAfterColumnC, '').getPropertyValue('grid-column-start') is "c"
PASS getComputedStyle(GridAreasNamedLineAfterColumnC, '').getPropertyValue('grid-column-end') is "c"
PASS getComputedStyle(GridAreasNamedLineAfterRowC, '').getPropertyValue('grid-column') is "auto / auto"
PASS getComputedStyle(GridAreasNamedLineAfterRowC, '').getPropertyValue('grid-column-start') is "auto"
PASS getComputedStyle(GridAreasNamedLineAfterRowC, '').getPropertyValue('grid-column-end') is "auto"
PASS getComputedStyle(GridAreasNamedLineAfterColumnStartAStart, '').getPropertyValue('grid-column') is "a-start / auto"
PASS getComputedStyle(GridAreasNamedLineAfterColumnStartAStart, '').getPropertyValue('grid-column-start') is "a-start"
PASS getComputedStyle(GridAreasNamedLineAfterColumnStartAStart, '').getPropertyValue('grid-column-end') is "auto"
PASS getComputedStyle(GridAreasNamedLineAfterRowStartCStart, '').getPropertyValue('grid-column') is "auto / auto"
PASS getComputedStyle(GridAreasNamedLineAfterRowStartCStart, '').getPropertyValue('grid-column-start') is "auto"
PASS getComputedStyle(GridAreasNamedLineAfterRowStartCStart, '').getPropertyValue('grid-column-end') is "auto"
PASS getComputedStyle(GridAreasNamedLineAfterRowStartCStart, '').getPropertyValue('grid-row') is "c-start / auto"
PASS getComputedStyle(GridAreasNamedLineAfterRowStartCStart, '').getPropertyValue('grid-row-start') is "c-start"
PASS getComputedStyle(GridAreasNamedLineAfterRowStartCStart, '').getPropertyValue('grid-row-end') is "auto"
PASS getComputedStyle(GridAreasNamedLineAfterColumnA, '').getPropertyValue('grid-column') is "a / a"
PASS getComputedStyle(GridAreasNamedLineAfterColumnA, '').getPropertyValue('grid-column-start') is "a"
PASS getComputedStyle(GridAreasNamedLineAfterColumnA, '').getPropertyValue('grid-column-end') is "a"
PASS getComputedStyle(GridAreasNamedLineAfterColumnARowD, '').getPropertyValue('grid-column') is "a / a"
PASS getComputedStyle(GridAreasNamedLineAfterColumnARowD, '').getPropertyValue('grid-column-start') is "a"
PASS getComputedStyle(GridAreasNamedLineAfterColumnARowD, '').getPropertyValue('grid-column-end') is "a"
PASS getComputedStyle(GridAreasNamedLineBeforeColumnD, '').getPropertyValue('grid-column') is "d / d"
PASS getComputedStyle(GridAreasNamedLineBeforeColumnD, '').getPropertyValue('grid-column-start') is "d"
PASS getComputedStyle(GridAreasNamedLineBeforeColumnD, '').getPropertyValue('grid-column-end') is "d"
PASS getComputedStyle(GridAreasNamedLineBeforeRowD, '').getPropertyValue('grid-column') is "auto / auto"
PASS getComputedStyle(GridAreasNamedLineBeforeRowD, '').getPropertyValue('grid-column-start') is "auto"
PASS getComputedStyle(GridAreasNamedLineBeforeRowD, '').getPropertyValue('grid-column-end') is "auto"
PASS getComputedStyle(GridAreasNamedLineBeforeColumnC, '').getPropertyValue('grid-column') is "c / c"
PASS getComputedStyle(GridAreasNamedLineBeforeColumnC, '').getPropertyValue('grid-column-start') is "c"
PASS getComputedStyle(GridAreasNamedLineBeforeColumnC, '').getPropertyValue('grid-column-end') is "c"
PASS getComputedStyle(GridAreasNamedLineBeforeRowC, '').getPropertyValue('grid-column') is "auto / auto"
PASS getComputedStyle(GridAreasNamedLineBeforeRowC, '').getPropertyValue('grid-column-start') is "auto"
PASS getComputedStyle(GridAreasNamedLineBeforeRowC, '').getPropertyValue('grid-column-end') is "auto"
PASS getComputedStyle(GridAreasNamedLineBeforeColumnStartDStart, '').getPropertyValue('grid-column') is "d-start / auto"
PASS getComputedStyle(GridAreasNamedLineBeforeColumnStartDStart, '').getPropertyValue('grid-column-start') is "d-start"
PASS getComputedStyle(GridAreasNamedLineBeforeColumnStartDStart, '').getPropertyValue('grid-column-end') is "auto"
PASS getComputedStyle(GridAreasNamedLineBeforeRowStartDStart, '').getPropertyValue('grid-column') is "auto / auto"
PASS getComputedStyle(GridAreasNamedLineBeforeRowStartDStart, '').getPropertyValue('grid-column-start') is "auto"
PASS getComputedStyle(GridAreasNamedLineBeforeRowStartDStart, '').getPropertyValue('grid-column-end') is "auto"
PASS getComputedStyle(GridAreasNamedLineBeforeRowStartDStart, '').getPropertyValue('grid-row') is "d-start / auto"
PASS getComputedStyle(GridAreasNamedLineBeforeRowStartDStart, '').getPropertyValue('grid-row-start') is "d-start"
PASS getComputedStyle(GridAreasNamedLineBeforeRowStartDStart, '').getPropertyValue('grid-row-end') is "auto"
PASS getComputedStyle(GridAreasNamedLineBeforeColumnA, '').getPropertyValue('grid-column') is "a / a"
PASS getComputedStyle(GridAreasNamedLineBeforeColumnA, '').getPropertyValue('grid-column-start') is "a"
PASS getComputedStyle(GridAreasNamedLineBeforeColumnA, '').getPropertyValue('grid-column-end') is "a"
PASS getComputedStyle(GridAreasNamedLineBeforeColumnARowD, '').getPropertyValue('grid-column') is "a / a"
PASS getComputedStyle(GridAreasNamedLineBeforeColumnARowD, '').getPropertyValue('grid-column-start') is "a"
PASS getComputedStyle(GridAreasNamedLineBeforeColumnARowD, '').getPropertyValue('grid-column-end') is "a"
PASS getComputedStyle(GridAreasNoNamesColumnAAStartRowDStart, '').getPropertyValue('grid-column') is "a / a-start"
PASS getComputedStyle(GridAreasNoNamesColumnAAStartRowDStart, '').getPropertyValue('grid-column-start') is "a"
PASS getComputedStyle(GridAreasNoNamesColumnAAStartRowDStart, '').getPropertyValue('grid-column-end') is "a-start"
PASS getComputedStyle(GridAreasNoNamesColumnDDStartRowCStart, '').getPropertyValue('grid-column') is "d / d-start"
PASS getComputedStyle(GridAreasNoNamesColumnDDStartRowCStart, '').getPropertyValue('grid-column-start') is "d"
PASS getComputedStyle(GridAreasNoNamesColumnDDStartRowCStart, '').getPropertyValue('grid-column-end') is "d-start"
PASS getComputedStyle(GridAreasNoNamesColumnCRowAAStart, '').getPropertyValue('grid-column') is "c / c"
PASS getComputedStyle(GridAreasNoNamesColumnCRowAAStart, '').getPropertyValue('grid-column-start') is "c"
PASS getComputedStyle(GridAreasNoNamesColumnCRowAAStart, '').getPropertyValue('grid-column-end') is "c"
PASS getComputedStyle(GridAreasNoNamesColumnCRowAAStart, '').getPropertyValue('grid-row') is "a / a-start"
PASS getComputedStyle(GridAreasNoNamesColumnCRowAAStart, '').getPropertyValue('grid-row-start') is "a"
PASS getComputedStyle(GridAreasNoNamesColumnCRowAAStart, '').getPropertyValue('grid-row-end') is "a-start"
PASS getComputedStyle(GridAreasNoNamesColumnDRowCCStart, '').getPropertyValue('grid-column') is "d / d"
PASS getComputedStyle(GridAreasNoNamesColumnDRowCCStart, '').getPropertyValue('grid-column-start') is "d"
PASS getComputedStyle(GridAreasNoNamesColumnDRowCCStart, '').getPropertyValue('grid-column-end') is "d"
PASS getComputedStyle(GridAreasNoNamesColumnDRowCCStart, '').getPropertyValue('grid-row') is "c / c-start"
PASS getComputedStyle(GridAreasNoNamesColumnDRowCCStart, '').getPropertyValue('grid-row-start') is "c"
PASS getComputedStyle(GridAreasNoNamesColumnDRowCCStart, '').getPropertyValue('grid-row-end') is "c-start"
PASS getComputedStyle(GridAreasEnd, '').getPropertyValue('grid-column') is "a / a"
PASS getComputedStyle(GridAreasEnd, '').getPropertyValue('grid-column-start') is "a"
PASS getComputedStyle(GridAreasEnd, '').getPropertyValue('grid-column-end') is "a"
PASS getComputedStyle(GridAreasEndColumnARowC, '').getPropertyValue('grid-column') is "a / a"
PASS getComputedStyle(GridAreasEndColumnARowC, '').getPropertyValue('grid-column-start') is "a"
PASS getComputedStyle(GridAreasEndColumnARowC, '').getPropertyValue('grid-column-end') is "a"
PASS getComputedStyle(GridAreasEndRowD, '').getPropertyValue('grid-column') is "auto / auto"
PASS getComputedStyle(GridAreasEndRowD, '').getPropertyValue('grid-column-start') is "auto"
PASS getComputedStyle(GridAreasEndRowD, '').getPropertyValue('grid-column-end') is "auto"
PASS getComputedStyle(GridAreasEndColumnARowD, '').getPropertyValue('grid-column') is "a / a"
PASS getComputedStyle(GridAreasEndColumnARowD, '').getPropertyValue('grid-column-start') is "a"
PASS getComputedStyle(GridAreasEndColumnARowD, '').getPropertyValue('grid-column-end') is "a"
PASS getComputedStyle(GridAreasRepeatColumnDRowC, '').getPropertyValue('grid-column') is "d / d"
PASS getComputedStyle(GridAreasRepeatColumnDRowC, '').getPropertyValue('grid-column-start') is "d"
PASS getComputedStyle(GridAreasRepeatColumnDRowC, '').getPropertyValue('grid-column-end') is "d"
PASS getComputedStyle(GridAreasRepeatColumnDStartDEndRowCStartCEnd, '').getPropertyValue('grid-column') is "d-start / d-end"
PASS getComputedStyle(GridAreasRepeatColumnDStartDEndRowCStartCEnd, '').getPropertyValue('grid-column-start') is "d-start"
PASS getComputedStyle(GridAreasRepeatColumnDStartDEndRowCStartCEnd, '').getPropertyValue('grid-column-end') is "d-end"
PASS getComputedStyle(GridAreasRepeatColumnDStartDEndRowCStartCEnd, '').getPropertyValue('grid-row') is "c-start / c-end"
PASS getComputedStyle(GridAreasRepeatColumnDStartDEndRowCStartCEnd, '').getPropertyValue('grid-row-start') is "c-start"
PASS getComputedStyle(GridAreasRepeatColumnDStartDEndRowCStartCEnd, '').getPropertyValue('grid-row-end') is "c-end"
PASS getComputedStyle(GridAreasRepeatColumnCRowD, '').getPropertyValue('grid-column') is "c / c"
PASS getComputedStyle(GridAreasRepeatColumnCRowD, '').getPropertyValue('grid-column-start') is "c"
PASS getComputedStyle(GridAreasRepeatColumnCRowD, '').getPropertyValue('grid-column-end') is "c"
PASS successfullyParsed is true

TEST COMPLETE
